add logic for opening member modal on click to card
